Add fetch to get local file while on web
Added the fetch to the viz.js file in order to have a way to get the staghunt.json file while online. Able to add the fetch and am not getting any errors but the visuals aren't showing. I think this might because we need to redefine (or just define) the "d3" object (180, 186) so that we're using the D3 library.