Basic Visuals
For example, I had to change the name from “index copy.html” to “index.html”  because that is the default name needed for it to run. Also added some of the missing fields of the html file and moved the script to the bottom of the file for reference reasons. I created another D3 folder and copied in the d3.min.js file because when hosting the files locally, using http-server, we can’t call files outside the root directory, which would be the example folder. Currently, the example should show a graph with random points on it.

For staghunt, I moved the script to the bottom of the file for referencing reasons. I used an http-server to test, which can be run with “run http-server”. Staghunt was undefined in the file so I implemented an asynchronous call using fetch to get the file we want. This can be changed and used at an IO between frontend and backend. As a result of the asynchronous call, I wrapped the character creation in a function that will execute once the file has been retrieved.