Huntspace Container Creation
Temporary save of the work done over the week of 01/02. Created a temporary dom to create the huntspace in and then export it to the main page. However, after much consideration this method seems to have two problems 1) many undefined variables now that many parts have now required explicit definition and 2) the animation of the transition between steps may not render with this method (which is vital). Going to look back at some D3 implementations and see what changes need to be made.